feat: control finder behavior by experimental feature flag
