+++ internal/features/finder.go
@@ -1,0 +1,5 @@
+//go:build viper_finder
+
+package features
+
+const Finder = true
