+++ internal/features/finder_default.go
@@ -1,0 +1,5 @@
+//go:build !viper_finder
+
+package features
+
+const Finder = false
